Feat/bulk node actions

+++ backend/src/services/nodes/Node.service.ts
@@ -1,5 +1,10 @@
 import { DB } from "@/config/db";
-import type { Node } from "@/domain/nodes/node";
+import type {
+  DirectoryNode,
+  FileNode,
+  Node,
+  NodeLite,
+} from "@/domain/nodes/node";
 import type { UploadedFile } from "@/domain/uploads/uploaded-file";
 import { isDirectoryNode } from "@/infra/guards/node";
 import type { Prisma } from "@/infra/prisma/generated/client";
@@ -17,8 +22,7 @@ import { CloudStorageService } from "../cloud/CloudStorage.service";
  * @description Servicio para gestionar nodos (archivos y directorios).
  */
 export class NodeService {
-  private static readonly persistence: NodePersistenceService =
-    new NodePersistenceService();
+  private static readonly persistence = NodePersistenceService;
   private static readonly cloud = CloudStorageService;
   private static readonly identity = NodeIdentityService;
   private static readonly repo = NodeRepository;
@@ -32,34 +36,36 @@ export class NodeService {
    */
   static async process(file: UploadedFile, parentId: string | null) {
     try {
-      // Resolver nombre y hash unicos
-      const { nodeName, nodeHash } = await this.identity.resolve(
-        file,
-        parentId,
-      );
+      return await this.prisma.$transaction(async (tx) => {
+        // Resolver nombre y hash unicos
+        const { nodeName, nodeHash } = await this.identity.resolveTx(
+          tx,
+          file,
+          parentId,
+        );
 
-      console.log(`Processing node: ${nodeName}`);
+        console.log(`Processing node: ${nodeName}`);
 
-      // Persistir el nodo en la base de datos
-      const node = await this.persistence.persist(
-        file,
-        parentId,
-        nodeName,
-        nodeHash,
-      );
+        // Persistir el nodo en la base de datos
+        const node = await this.persistence.persistTx(
+          tx,
+          file,
+          parentId,
+          nodeName,
+          nodeHash,
+        );
 
-      // Actualizar el tamaño del nodo padre si es una carpeta
-      await this.prisma.$transaction(async (tx) => {
+        // Actualizar el tamaño del nodo padre si es una carpeta
         if (parentId) {
           const parentNode = await this.repo.findByIdTx(tx, parentId);
           if (parentNode) {
             await this.incrementNodeSizeByIdTx(tx, parentId, file.size);
           }
         }
-      });
 
-      console.log(`Node processed: ${nodeName} as ${nodeHash}`);
-      return node;
+        console.log(`Node processed: ${nodeName} as ${nodeHash}`);
+        return node;
+      });
     } catch (err) {
       console.error(err);
       await this.cloud.delete(file.path); // Limpiar archivo temporal en caso de error
@@ -97,6 +103,25 @@ export class NodeService {
   }
 
   /**
+   * @description Obtiene los detalles de múltiples nodos por sus IDs.
+   * @param nodeIds Array de IDs de los nodos a obtener
+   * @returns Array de nodos con sus detalles
+   */
+  static async getNodesDetailsBulk(nodeIds: Node["id"][]): Promise<Node[]> {
+    try {
+      const nodes = await this.repo.findManyByIds(nodeIds);
+      return nodes;
+    } catch (err) {
+      if (err instanceof AppError) throw err;
+      else
+        throw new AppError(
+          "INTERNAL",
+          `Error al obtener los detalles de los nodos`,
+        );
+    }
+  }
+
+  /**
    * @description Obtiene todos los ancestros de un nodo dado.
    * @param startNodeId ID del nodo desde el cual comenzar a buscar ancestros
    * @returns Array de ancestros del nodo
@@ -128,25 +153,29 @@ export class NodeService {
     name: string | null,
   ): Promise<Omit<Node, "hash">> {
     try {
-      // Si no hay nombre, colocamos uno por defecto
-      let finalName = name ?? "Untitled Folder";
+      return await this.prisma.$transaction(async (tx) => {
+        // Si no hay nombre, colocamos uno por defecto
+        let finalName = name ?? "Untitled Folder";
 
-      // Verificamos que no exista un directorio con el mismo nombre
-      const existentNode = await this.repo.findDirByName(parentId, finalName);
+        // Verificamos que no exista un directorio con el mismo nombre
+        const existentNode = await this.repo.findDirByNameTx(
+          tx,
+          parentId,
+          finalName,
+        );
 
-      // En caso de que exista un directorio, resolvemos el nombre
-      if (existentNode) {
-        finalName = await this.resolveName(parentId, finalName);
-      }
+        // En caso de que exista un directorio, resolvemos el nombre
+        if (existentNode) {
+          finalName = await this.resolveName(parentId, finalName);
+        }
 
-      // Preparamos los datos para crear el directorio
-      const hash = NodeUtils.genDirectoryHash(finalName, parentId);
-      const mime = "inode/directory";
+        // Preparamos los datos para crear el directorio
+        const hash = NodeUtils.genDirectoryHash(finalName, parentId);
+        const mime = "inode/directory";
 
-      // Tratamos de crear el directorio (nodo al fin)
+        // Tratamos de crear el directorio (nodo al fin)
 
-      if (parentId) {
-        return await this.prisma.$transaction(async (tx) => {
+        if (parentId) {
           const { hash: _h, ...node } = await this.repo.createTx(tx, {
             name: finalName,
             hash,
@@ -167,21 +196,21 @@ export class NodeService {
 
           // Si sale bien, devolvemos el directorio creado en el parentId correspondiente
           return node;
+        }
+
+        // Crear el directorio sin padre (raiz)
+        const { hash: _h, ...node } = await this.repo.createTx(tx, {
+          name: finalName,
+          hash,
+          parent: undefined,
+          size: 0,
+          mime,
+          isDir: true,
         });
-      }
 
-      // Crear el directorio sin padre (raiz)
-      const { hash: _h, ...node } = await this.createNode({
-        name: finalName,
-        hash,
-        parent: undefined,
-        size: 0,
-        mime,
-        isDir: true,
+        // Si sale bien, devolvemos el directorio creado en el parentId correspondiente
+        return node;
       });
-
-      // Si sale bien, devolvemos el directorio creado en el parentId correspondiente
-      return node;
     } catch (err) {
       console.error(err);
       throw new AppError("INTERNAL", "Error al crear el directorio (nodo)");
@@ -198,17 +227,20 @@ export class NodeService {
     // Asegurarse de que la extension del nodo se mantenga igual
     newName = NodeUtils.ensureNodeExt(newName, node);
 
-    // Resolver nombre y hash unicos
-    const { nodeName, nodeHash } = await this.identity.resolve(
-      node,
-      node.parentId,
-      { newName },
-    );
+    return await this.prisma.$transaction(async (tx) => {
+      // Resolver nombre y hash unicos
+      const { nodeName, nodeHash } = await this.identity.resolveTx(
+        tx,
+        node,
+        node.parentId,
+        { newName },
+      );
 
-    // Actualizar el nodo en la base de datos
-    return await this.repo.updateIdentityById(node.id, {
-      newName: nodeName,
-      newHash: nodeHash,
+      // Actualizar el nodo en la base de datos
+      return await this.repo.updateIdentityByIdTx(tx, node.id, {
+        newName: nodeName,
+        newHash: nodeHash,
+      });
     });
   }
 
@@ -277,13 +309,12 @@ export class NodeService {
     node: Node,
     parentId: string | null,
     newName?: string,
-  ): Promise<Node | Node[]> {
+  ): Promise<NodeLite | NodeLite[]> {
     try {
       // Copiar el nuevo nodo de forma física y añadir un nueva fila a la base de datos
       if (isDirectoryNode(node)) {
         return await NodeTreeService.copyNodeDir(node, parentId, {
           newName,
-          concurrency: 5,
         });
       } else {
         return await NodeTreeService.attachNodeFile(node, parentId, newName);
@@ -295,6 +326,42 @@ export class NodeService {
       } else {
         throw new AppError("INTERNAL", "Error al copiar el nodo");
       }
+    }
+  }
+
+  /**
+   * @description Copia varios nodos (archivos y directorios) a una nueva ubicación.
+   * @param nodes Nodos a copiar
+   * @param parentId ID del nodo padre donde se ubicará las copias
+   * @returns Nodos copiados
+   */
+  static async bulkCopyNodes(
+    nodes: Node[],
+    parentId: Node["parentId"],
+  ): Promise<NodeLite[]> {
+    try {
+      const copiedNodes: NodeLite[] = [];
+      const directories = nodes.filter((n) => n.isDir);
+      const files = nodes.filter((n) => !n.isDir);
+
+      if (directories.length > 0) {
+        copiedNodes.push(
+          ...(await NodeTreeService.bulkCopyNodeDirs(directories, parentId)),
+        );
+      }
+
+      if (files.length > 0) {
+        copiedNodes.push(
+          ...(await NodeTreeService.bulkAttachNodeFiles(files, parentId)),
+        );
+      }
+
+      return copiedNodes;
+    } catch (err) {
+      console.error(err);
+      if (err instanceof AppError) throw err;
+      else
+        throw new AppError("INTERNAL", "No se pudieron copiar uno o más nodos");
     }
   }
 
@@ -320,7 +387,6 @@ export class NodeService {
       if (isDirectoryNode(node)) {
         return await NodeTreeService.moveNodeDir(node, parentId, {
           newName,
-          concurrency: 5,
         });
       } else {
         return await NodeTreeService.moveNodeFile(node, parentId, newName);
@@ -336,13 +402,64 @@ export class NodeService {
   }
 
   /**
-   * @description Elimina un nodo.
+   * @description Mueve varios nodos (archivos y directorios) a una nueva ubicación.
+   * @param nodes Nodos a mover
+   * @param parentId ID del nodo padre donde se ubicará los nodos movidos
+   * @returns Nodos movidos
+   */
+  static async bulkMoveNodes(
+    nodes: Node[],
+    parentId: Node["parentId"],
+  ): Promise<NodeLite[]> {
+    try {
+      const movedNodes: NodeLite[] = [];
+      const directories = nodes.filter((n) => n.isDir);
+      const files = nodes.filter((n) => !n.isDir);
+
+      if (directories.length > 0) {
+        movedNodes.push(
+          ...(await NodeTreeService.bulkMoveNodeDirs(directories, parentId)),
+        );
+      }
+
+      if (files.length > 0) {
+        movedNodes.push(
+          ...(await NodeTreeService.bulkMoveNodeFiles(files, parentId)),
+        );
+      }
+
+      return movedNodes;
+    } catch (err) {
+      console.error(err);
+      if (err instanceof AppError) throw err;
+      else
+        throw new AppError("INTERNAL", "No se pudieron mover uno o más nodos");
+    }
+  }
+
+  /**
+   * @description Elimina un nodo (archivo o directorio).
    * @param node Nodo a eliminar
    */
   static async deleteNode(node: Node) {
+    if (node.isDir) {
+      await NodeService.deleteDirectory(node);
+    } else {
+      await NodeService.deleteFileNode(node);
+    }
+  }
+
+  /**
+   * @description Elimina un nodo.
+   * @param node Nodo a eliminar
+   */
+  static async deleteFileNode(node: FileNode) {
+    // Obtener la ruta del nodo
+    const nodePath = this.cloud.getFilePath(node);
+
     try {
-      // Obtener la ruta del nodo
-      const nodePath = this.cloud.getFilePath(node);
+      // Eliminar el nodo del sistema de nodos
+      await this.cloud.delete(nodePath);
 
       // Usar transacción para eliminar el nodo y actualizar tamaños
       await this.prisma.$transaction(async (tx) => {
@@ -359,12 +476,54 @@ export class NodeService {
         // Eliminar el registro del nodo en la base de datos
         await this.repo.deleteByIdTx(tx, node.id);
       });
-
-      // Eliminar el nodo del sistema de nodos
-      await this.cloud.delete(nodePath);
     } catch (err) {
       if (err instanceof AppError) throw err;
       else throw new AppError("INTERNAL", "Error al eliminar el nodo");
+    }
+  }
+
+  /**
+   * @description Elimina varios nodos.
+   * @param nodes Nodos a eliminar
+   * @returns Promise<void>
+   */
+  static async bulkDeleteFileNodes(nodes: FileNode[]) {
+    // Obtener las rutas de los nodos
+    const nodePaths = nodes.map((n) => this.cloud.getFilePath(n));
+
+    try {
+      // Primer borrar fisico por si lanza error atraparlo antes de tocar la base de datos
+      // Eliminar los archivos del sistema de nodos
+      for (const path of nodePaths) {
+        await this.cloud.delete(path);
+      }
+
+      return await this.prisma.$transaction(
+        async (tx) => {
+          for (const node of nodes) {
+            // Si tiene padre, actualizar el tamaño de todos los ancestros que haya
+            if (node.parentId) {
+              const parent = await this.repo.findByIdTx(tx, node.parentId);
+
+              if (parent) {
+                // Actualizar el tamaño de todos los ancestros
+                await this.decrementNodeSizeByIdTx(
+                  tx,
+                  node.parentId,
+                  node.size,
+                );
+              }
+            }
+
+            // Eliminar el registro del nodo en la base de datos
+            await this.repo.deleteByIdTx(tx, node.id);
+          }
+        },
+        { maxWait: 5000, timeout: 60000 }, // 60 segundos de timeout por si hay muchos nodos
+      );
+    } catch (err) {
+      if (err instanceof AppError) throw err;
+      else throw new AppError("INTERNAL", "Error al eliminar los nodos");
     }
   }
 
@@ -377,9 +536,9 @@ export class NodeService {
     try {
       // Eliminar los archivos creados en el sistema de nodos
       const tmpPaths = uploadedFiles.map((f) => f.path);
-      const nodePaths = createdNodes.map((n) =>
-        CloudStorageService.getFilePath(n),
-      );
+      const nodePaths = createdNodes
+        .filter((n) => !n.isDir)
+        .map((n) => CloudStorageService.getFilePath(n));
       await CloudStorageService.deleteFiles([...tmpPaths, ...nodePaths]);
 
       // Eliminar los nodos ya creados en la base de datos
@@ -396,22 +555,26 @@ export class NodeService {
    * @description Elimina un nodo de tipo directorio y todos sus nodos descendientes.
    * @param node Nodo de tipo directorio a eliminar junto con todos sus descendientes.
    */
-  static async deleteDirectory(node: Node) {
+  static async deleteDirectory(node: DirectoryNode) {
     try {
-      if (!node.isDir) {
-        throw new AppError("NODE_IS_NOT_DIRECTORY");
-      }
-
       // Obtenemos los descendientes de la carpeta (incluyéndola)
       const descendants = await this.repo.getAllNodeDescendants(node.id);
 
       // Si la carpeta está vacía (no contiene archivos ni subdirectorios).
-      if (descendants.length == 0) {
+      if (descendants.length == 1 && descendants[0].id === node.id) {
         await this.prisma.$transaction(async (tx) => {
           // Eliminar el registro del nodo en la base de datos
           await this.repo.deleteByIdTx(tx, node.id);
         });
         return;
+      }
+
+      // Eliminar los archivos de la nube
+      // Ahora iterando para evitar borrar todos en caso de error en uno solo
+      for (const d of descendants) {
+        if (!d.isDir) {
+          await this.cloud.delete(this.cloud.getFilePath(d));
+        }
       }
 
       // Usar transacción para actualizar los tamaños
@@ -432,16 +595,6 @@ export class NodeService {
           descendants.map((descendant) => descendant.id),
         );
       });
-
-      // Obtenemos las rutas de los archivos descendientes
-      const nodePaths = descendants
-        .filter((descendant) => !descendant.isDir)
-        .map((descendant) => this.cloud.getFilePath(descendant));
-
-      if (nodePaths) {
-        // Eliminamos los archivos
-        await this.cloud.deleteFiles(nodePaths);
-      }
     } catch (err) {
       if (err instanceof AppError) throw err;
       else throw new AppError("INTERNAL", "Error al eliminar el nodo");
@@ -449,16 +602,42 @@ export class NodeService {
   }
 
   /**
-   * @description Actualiza el nombre de un nodo.
-   * @param node Nodo a actualizar
-   * @param newName Nuevo nombre para el nodo
-   * @returns Nodo actualizado
+   * @description Elimina varios nodos de tipo directorio y todos sus nodos descendientes.
+   * @param nodes Nodos de tipo directorio a eliminar junto con todos sus descendientes.
    */
-  static async updateNodeName(
-    nodeId: Node["id"],
-    newName: string,
-  ): Promise<Node> {
-    return await this.repo.updateNameById(nodeId, newName);
+  static async bulkDeleteDirectories(nodes: DirectoryNode[]) {
+    try {
+      // Fuera mas optimizado de otra forma pero esto iterando asi mantenemos la consistencia
+      // Y evitamos problemas a futuro ya que el borrado es una accion muy destructiva
+      // Es mejor tener control total sobre cada eliminacion
+
+      // Iterar sobre todos los nodos a eliminar
+      for (const node of nodes) {
+        await this.deleteDirectory(node);
+      }
+    } catch (err) {
+      if (err instanceof AppError) throw err;
+      else throw new AppError("INTERNAL", "Error al eliminar los nodos");
+    }
+  }
+
+  /**
+   * @description Elimina varios nodos (archivos y directorios).
+   * @param nodes Nodos a eliminar
+   */
+  static async bulkDeleteNodes(nodes: Node[]) {
+    const fileNodes = nodes.filter((n) => !n.isDir);
+    const dirNodes = nodes.filter((n) => n.isDir);
+
+    // Eliminar archivos primero
+    if (fileNodes.length > 0) {
+      await this.bulkDeleteFileNodes(fileNodes);
+    }
+
+    // Luego eliminar directorios
+    if (dirNodes.length > 0) {
+      await this.bulkDeleteDirectories(dirNodes);
+    }
   }
 
   /**
